+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix `audio_get_status` result
+- Add adapter auto detection for unspecified port
+- Fix `is_active_source` result
 - Fix calling `libcec_close` twice on drop
 
 ## 7.1.1